Adapt tests to framework refactor
