+++ tests/integration/test_remoted/conftest.py
@@ -27,12 +27,12 @@ def restart_wazuh_expect_error() -> None:
     services.control_service('stop')
 
 
-'''
-    description: gets 'elements' configuration parameters  from 'sections' field and converts it from list to list dict
-    return  configuration parameters
-'''
 @pytest.fixture
 def get_real_configuration(test_configuration):
+    """
+        description: gets 'elements' configuration parameters  from 'sections' field and converts it from list to list dict
+        return  configuration parameters
+    """
     config_data = test_configuration.get('sections', {})[0]['elements']
     real_config = dict()
 
@@ -46,3 +46,28 @@ def get_real_configuration(test_configuration):
     real_config_list = list()
     real_config_list.append(real_config)
     return real_config_list
+
+
+@pytest.fixture
+def protocols_list_to_str_upper_case(request, test_metadata):
+    """convert valid_protocol list to comma separated uppercase string
+
+        parameters: test_metadata
+        Returns:
+        protocol_string_upper: string protocols in uppercase
+    """
+    protocol_array = []
+    protocol_string_upper = []
+    for I in test_metadata['valid_protocol'] :
+        protocol_array.append(I)
+        protocol_array.sort()
+
+    if len(test_metadata['valid_protocol']) > 0 :
+        protocol_string = protocol_array[0]
+        protocol_string_upper = protocol_string.upper();
+
+    if len(protocol_array) > 1:
+        protocol_string = protocol_array[0] + ',' + protocol_array[1]
+        protocol_string_upper = protocol_string.upper();
+
+    return protocol_string_upper
