+++ tests/integration/test_remoted/test_socket_communication/test_syslog_message.py
@@ -13,7 +13,6 @@ from wazuh_testing.utils.callbacks import generate_callback
 from wazuh_testing.utils.configuration import get_test_cases_data, load_configuration_template
 from wazuh_testing.constants.paths.logs import ARCHIVES_LOG_PATH
 from wazuh_testing.modules.remoted.configuration import REMOTED_DEBUG
-from wazuh_testing.scripts.syslog_simulator import SYSLOG_SIMULATOR_START_TIME
 from wazuh_testing.tools import thread_executor
 from wazuh_testing.tools.simulators import run_syslog_simulator
 
@@ -33,6 +32,8 @@ daemons_handler_configuration = {'all_daemons': True}
 
 local_internal_options = {REMOTED_DEBUG: '2'}
 
+# Test variables.
+SYSLOG_SIMULATOR_START_TIME = 2
 
 # Test function.
 @pytest.mark.parametrize('test_configuration, test_metadata',  zip(test_configuration, test_metadata), ids=cases_ids)
